requirements.txt
beautifulsoup4==4.7.1
dialogflow==0.5.2
requests==2.20.1
slackclient==1.3.0
urllib3==1.22
